Changes to homework folder
originally i added the hw1 contents to hw folder and pushed because i did something wrong. Now when i removed and added back the hw1 folder from the homework folder it says no changes. Therefore I removed the homework folder and took a screenshot for the before and will now be pushing to main to illustrate the change.